ca plane pour moi

+++ srcs/vector_array/vector_utils.c
@@ -6,7 +6,7 @@
 /*   By: tshimoda <tshimoda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/29 12:04:58 by tshimoda          #+#    #+#             */
-/*   Updated: 2022/02/25 14:11:26 by tshimoda         ###   ########.fr       */
+/*   Updated: 2022/03/10 20:39:35 by tshimoda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	vector_free_elements(t_vector *v)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (v->elements[i])
